% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_code.R
\name{avg_subset_parallel}
\alias{avg_subset_parallel}
\title{subset metrics and average across shifts}
\usage{
avg_subset_parallel(x, shift, metrics, na.rm, wopt)
}
\arguments{
\item{x}{a packed SpatRaster}

\item{shift}{a list of shifts from glcm_textures}

\item{metrics}{Requested GLCM textures metrics to be returned}

\item{na.rm}{logical indicating if NA values should be removed from calculations}

\item{wopt}{Additional options}
}
\value{
a packed SpatRaster
}
\description{
subset metrics and average across shifts
}
