% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glcm_metrics.R
\name{glcm_metrics}
\alias{glcm_metrics}
\title{Calculates the GLCM Texture Metrics from a GLCM}
\usage{
glcm_metrics(GLCM)
}
\arguments{
\item{GLCM}{A numeric matrix representing a Normalized GLCM}
}
\description{
Calculates the GLCM Texture Metrics from a GLCM
}
\references{
Hall-Beyer, M., 2017. GLCM Textrure: A Tutorial v. 3.0. University of Calgary, Alberta, Canada.

Haralick, R.M., Shanmugam, K., Dinstein, I., 1973. Textural features for image classification. IEEE Transactions on Systems, Man, and Cybernetics 610–621. https://doi.org/10.1109/TSMC.1973.4309314
}
