% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glcm_textures.R
\name{glcm_textures}
\alias{glcm_textures}
\title{Calculates GLCM texture metrics of a Raster Layer}
\usage{
glcm_textures(
  r,
  w,
  n_levels,
  shift = list(c(1, 0), c(1, 1), c(0, 1), c(-1, 1)),
  metrics = c("glcm_contrast", "glcm_dissimilarity", "glcm_homogeneity", "glcm_ASM",
    "glcm_entropy", "glcm_mean", "glcm_variance", "glcm_correlation"),
  quantization,
  min_val = NULL,
  max_val = NULL
)
}
\arguments{
\item{r}{A raster layer. If already quantized set quantization to "none". The valid range of values for a quantized raster is from 0 to n_levels-1 (e.g. a raster with 32 grey levels would have a valid range of 0-31).}

\item{w}{A vector of length 2 specifying the dimensions of the rectangular window to use where the first number is the number of rows and the second number is the number of columns. Window size must be an odd number.}

\item{n_levels}{Number of grey levels used in the quantization}

\item{shift}{A vector of length 2, or a list of vectors each of length 2 specifying the relationship between neighboring pixel to the reference pixel. The first number represents the shift in the x direction and the second number represents the shift in the y direction, where up and right are positive. For example c(1,0) is the pixel directly to the right. The GLCM is made symmetrical by counting each pair twice, once "forwards" and once "backwards" by interchanging reference and neighbor pixels. Therefore a shift directly to the right c(1,0) is equivalent to a shift directly to the left c(-1,0). To average over "all directions" you can use shift=list(c(1,0), c(1,1), c(0,1), c(-1,1)), which is the default.}

\item{metrics}{A vector of glcm texture metrics to return. Valid entries include "glcm_contrast", "glcm_dissimilarity", "glcm_homogeneity", "glcm_ASM", "glcm_entropy", "glcm_mean", "glcm_variance", "glcm_correlation".}

\item{quantization}{quantization method (either "equal range" or "equal prob"). "equal range" quantization will create bins that cover a range of equal size. "equal prob" performs equal probability quantization and will use quantiles to create bins with approximately equal number of samples.}

\item{min_val}{minimum value for equal range quantization (if not supplied, the minimum value of the raster is used)}

\item{max_val}{maximum value for equal range quantization (if not supplied, the maximum value of the raster is used)}
}
\value{
a RasterBrick of Texture Metrics (or RasterLayer if just one metric is calculated)
}
\description{
Calculates GLCM texture metrics of a RasterLayer over a sliding rectangular window
}
