% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_glcm.R
\name{make_glcm}
\alias{make_glcm}
\title{Creates a symmetrical normalized GLCM for a given matrix and shift}
\usage{
make_glcm(x, n_levels, shift)
}
\arguments{
\item{x}{a matrix of integers representing quantized values. The valid range of values is from 0 to n_levels-1 (e.g. a matrix with 32 grey levels would have a valid range of 0-31).}

\item{n_levels}{Number of grey levels used in the quantization}

\item{shift}{A vector of length 2 specifying the relationship between neighboring pixel to the reference pixel. The first number represents the shift in the x direction and the second number represents the shift in the y direction, where up and right are positive. For example c(1,0) is the pixel directly to the right. The GLCM is made symmetrical by counting each pair twice, once "forwards" and once "backwards" by interchanging reference and neighbor pixels. Therefore a shift directly to the right c(1,0) is equivalent to a shift directly to the left c(-1,0)}
}
\description{
Creates a symmetrical normalized GLCM for a given matrix and shift
}
