% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mat.R
\docType{data}
\name{test_mat}
\alias{test_mat}
\title{Test Matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2844 rows and 1126 columns.
}
\usage{
test_mat
}
\description{
Test Matrix
}
\keyword{datasets}
