% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glcm_metrics.R
\name{glcm_metrics}
\alias{glcm_metrics}
\title{Calculates the GLCM Texture Metrics from a GLCM}
\usage{
glcm_metrics(GLCM)
}
\arguments{
\item{GLCM}{A numeric matrix representing a Normalized GLCM}
}
\description{
Calculates the GLCM Texture Metrics from a GLCM
}
